refactor: split awake.py command handlers into dedicated module
Extract ~420 lines of /command handlers from awake.py (1045→522 lines)
into command_handlers.py. Shared state (KOAN_ROOT, INSTANCE_DIR, etc.)
lives in bridge_state.py to avoid circular imports.

awake.py retains: main loop, chat, outbox, message classification.
command_handlers.py owns: all /command handlers (help, status, usage,
log, mission, pause, resume, stop, verbose, silent, sparring, reflect,
ping). Backward-compatible re-exports in awake.py preserve the public API.

All 937 tests pass with updated patch targets.
